Throw InvalidProgramException
