+++ src/coreclr/src/tools/Common/TypeSystem/Common/ExceptionStringID.cs
@@ -36,6 +36,7 @@ namespace Internal.TypeSystem
         InvalidProgramNonStaticMethod,
         InvalidProgramGenericMethod,
         InvalidProgramNonBlittableTypes,
+        InvalidProgramMultipleCallConv,
 
         // BadImageFormatException
         BadImageFormatGeneric,
