--- src/tests/baseservices/callconvs/CallFunctionPointersNotSupported.ilproj
@@ -1,8 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.IL">
-  <PropertyGroup>
-    <OutputType>library</OutputType>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="CallFunctionPointersNotSupported.il" />
-  </ItemGroup>
-</Project>
