--- src/tests/baseservices/callconvs/TestCallingConventionsNotSupported.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <CrossGenTest>false</CrossGenTest>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="NativeFunctions.cs" />
-    <Compile Include="TestCallingConventionsNotSupported.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="CMakeLists.txt" />
-    <ProjectReference Include="CallFunctionPointersNotSupported.ilproj" />
-    <ProjectReference Include="$(TestSourceDir)Common/CoreCLRTestLibrary/CoreCLRTestLibrary.csproj" />
-  </ItemGroup>
-</Project>
